Refactor Account page into a stateless component
